requirements.txt
aiofiles==24.1.0
alembic==1.14.0
annotated-types==0.7.0
asgiref==3.8.1
asyncpg==0.30.0
blinker==1.9.0
boto3==1.35.93
botocore==1.35.93
click==8.1.7
dependency-injector==4.44.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.17.3
hyperframe==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
Mako==1.3.8
MarkupSafe==3.0.2
marshmallow==3.23.1
marshmallow-sqlalchemy==1.1.0
mypy-boto3-s3==1.35.93
packaging==24.2
priority==2.0.0
psycopg2==2.9.10
pydantic==2.10.3
pydantic_core==2.27.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
Quart==0.19.9
quart-cors==0.8.0
s3transfer==0.10.4
six==1.17.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.1.3
wsproto==1.2.0
